Add new Jupyter Notebook for Flourish Project
The new "redmine_7306.ipynb" file is added to Flourish Notebooks folder. This notebook is created to modify and process caregiver and child data through Django shell. It provides an interactive way of managing the CaregiverChildConsent data model by assigning names to the children and saving them directly in the database, to better handle and track multiple children associated with each caregiver. This solution was needed to avoid the manual updating of fields and improve data management and efficiency.