:hammer: ci(Script): "Replaced all_screenings static data with dynamic data from json file"
The code was updated to replace the static 'all_screenings' data with dynamic data read from a local JSON file 'screeningpregwomen.json'. This change was made to allow the flexibility of updating screening data without needing to modify the source code. The JSON file is read and parsed into a Python list, from which the all_screenings variable is populated.

Additionally, the attributes within the screening object were updated to fetch data from the parsed JSON data. The creation or update of ScreeningPregWomenInline objects was wrapped in a try-except block to handle the case when the corresponding ScreeningPregWomen object does not exist.