Add additional fields in model classes and handle exceptions
Added additional fields in the CaregiverChildConsent class and handled exceptions for the ImproperlyConfigured error. Modified lines of code to accept first name and last name, and updated the RegisteredSubject and ChildBirth objects. Changes are done to better capture caregiver's child consent in the 'flourish_caregiver' dataset and improve data quality.